+++ src/common/lib.rs
@@ -6,7 +6,6 @@
 #[macro_use]
 extern crate bitflags;
 extern crate memmap;
-extern crate rustc_serialize;
 extern crate typed_arena;
 
 #[macro_use]